File layout target - refactor 2
